horas fix

+++ src/proyectooperativos/CajaRegistradora.java
@@ -57,11 +57,11 @@ public class CajaRegistradora extends Thread{
                        System.out.println("El cajero #" + this.id + " está atendiendo a cliente #" + idCliente );
                        for(int i = 0; i < App.gama.mostradores.get(this.id).productos.size(); i++){
                                 System.out.println("El cajero #" + this.id + " pasó el producto #" + App.gama.mostradores.get(this.id).productos.get(i).getId() + " del cliente #" + idCliente );
-                                Thread.sleep( ( (1/60) *App.duracionDeHora) * 1000 );
+                                Thread.sleep( (long) ( ( (1f/60f) *App.duracionDeHora) * 1000 ) );
                                 ganancias = ganancias + App.gama.mostradores.get(this.id).productos.get(i).getPrecio();
                        }
                        
-                      Thread.sleep( ( ((1/2)/60) *App.duracionDeHora) * 1000 ); //0.5 min;
+                      Thread.sleep( (long) ( ( ((1f/2f)/60f) *App.duracionDeHora) ) * 1000 ); //0.5 min;
                        System.out.println("Agregadas " + ganancias + " a las ganancias totales");
                        App.gananciasTotales = App.gananciasTotales + ganancias;
                        App.gama.mostradores.get(this.id).clear();
